将bottom设置部分单独放在一个view controller中，该用继承view controller的方式添加bottom bar；在BottomBarViewController中添加几个跟高度计算相关的方法

+++ iosapp/OSCUser.h
@@ -13,9 +13,9 @@
 @property (readonly, nonatomic, assign) int64_t userID;
 @property (readonly, nonatomic, strong) NSString *location;
 @property (readonly, nonatomic, strong) NSString *name;
-@property (readonly, nonatomic, assign) unsigned long followersCount;
-@property (readonly, nonatomic, assign) unsigned long fansCount;
-@property (readonly, nonatomic, assign) long score;
+@property (readonly, nonatomic, assign) NSUInteger followersCount;
+@property (readonly, nonatomic, assign) NSUInteger fansCount;
+@property (readonly, nonatomic, assign) NSInteger score;
 @property (readonly, nonatomic, copy) NSURL *portraitURL;
 
 @end
